+++ src/main/java/com/dopamingu/be/domain/auth/application/AuthService.java
@@ -9,9 +9,9 @@ import com.dopamingu.be.domain.auth.dto.TokenRefreshRequest;
 import com.dopamingu.be.domain.global.error.exception.CustomException;
 import com.dopamingu.be.domain.global.error.exception.ErrorCode;
 import com.dopamingu.be.domain.global.util.MemberUtil;
-import com.dopamingu.be.domain.member.dao.MemberRepository;
 import com.dopamingu.be.domain.member.domain.Member;
 import com.dopamingu.be.domain.member.domain.OauthInfo;
+import com.dopamingu.be.domain.member.repository.MemberRepository;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
