+++ src/main/java/com/dopamingu/be/domain/member/application/MemberService.java
@@ -3,10 +3,10 @@ package com.dopamingu.be.domain.member.application;
 import com.dopamingu.be.domain.global.error.exception.CustomException;
 import com.dopamingu.be.domain.global.error.exception.ErrorCode;
 import com.dopamingu.be.domain.global.util.MemberUtil;
-import com.dopamingu.be.domain.member.dao.MemberRepository;
 import com.dopamingu.be.domain.member.domain.Member;
 import com.dopamingu.be.domain.member.dto.response.UsernameAvailableResponseDto;
 import com.dopamingu.be.domain.member.dto.response.UsernameResponseDto;
+import com.dopamingu.be.domain.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
