+++ src/test/java/com/dopamingu/be/domain/member/application/MemberServiceTest.java
@@ -9,13 +9,13 @@ import static org.mockito.Mockito.when;
 import com.dopamingu.be.domain.global.error.exception.CustomException;
 import com.dopamingu.be.domain.global.error.exception.ErrorCode;
 import com.dopamingu.be.domain.global.util.MemberUtil;
-import com.dopamingu.be.domain.member.dao.MemberRepository;
 import com.dopamingu.be.domain.member.domain.Member;
 import com.dopamingu.be.domain.member.domain.MemberStatus;
 import com.dopamingu.be.domain.member.domain.OauthInfo;
 import com.dopamingu.be.domain.member.domain.OauthProvider;
 import com.dopamingu.be.domain.member.dto.response.UsernameAvailableResponseDto;
 import com.dopamingu.be.domain.member.dto.response.UsernameResponseDto;
+import com.dopamingu.be.domain.member.repository.MemberRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
